Fixed minor compiler warnings

+++ test/unit/queue/fcqueue.cpp
@@ -152,16 +152,16 @@ namespace {
             // enqueue/dequeue
             for ( int i = 0; i < nSize; ++i ) {
                 ASSERT_TRUE( q.enqueue( value_type( i )));
-                ASSERT_EQ( q.size(), i + 1 );
+                ASSERT_EQ( q.size(), static_cast<size_t>( i + 1 ));
             }
             ASSERT_FALSE( q.empty() );
-            ASSERT_EQ( q.size(), nSize );
+            ASSERT_EQ( q.size(), static_cast<size_t>( nSize ));
 
             for ( int i = 0; i < nSize; ++i ) {
                 it.value = -1;
                 ASSERT_TRUE( q.dequeue( it ) );
                 ASSERT_EQ( it.value, i );
-                ASSERT_EQ( q.size(), nSize - i - 1 );
+                ASSERT_EQ( q.size(), static_cast<size_t>( nSize - i - 1 ));
             }
             ASSERT_TRUE( q.empty() );
             ASSERT_EQ( q.size(), 0u );
@@ -169,16 +169,16 @@ namespace {
             // push/pop
             for ( int i = 0; i < nSize; ++i ) {
                 ASSERT_TRUE( q.push( value_type( i )));
-                ASSERT_EQ( q.size(), i + 1 );
+                ASSERT_EQ( q.size(), static_cast<size_t>( i + 1 ));
             }
             ASSERT_FALSE( q.empty() );
-            ASSERT_EQ( q.size(), nSize );
+            ASSERT_EQ( q.size(), static_cast<size_t>( nSize ));
 
             for ( int i = 0; i < nSize; ++i ) {
                 it.value = -1;
                 ASSERT_TRUE( q.pop( it ) );
                 ASSERT_EQ( it.value, i );
-                ASSERT_EQ( q.size(), nSize - i - 1 );
+                ASSERT_EQ( q.size(), static_cast<size_t>( nSize - i - 1 ));
             }
             ASSERT_TRUE( q.empty() );
             ASSERT_EQ( q.size(), 0u );
@@ -188,7 +188,7 @@ namespace {
                 ASSERT_TRUE( q.push( value_type( i )));
             }
             ASSERT_FALSE( q.empty() );
-            ASSERT_EQ( q.size(), nSize );
+            ASSERT_EQ( q.size(), static_cast<size_t>( nSize ));
 
             q.clear();
             ASSERT_TRUE( q.empty() );
